대시보드_중간/requirements.txt
# ── 핵심 프레임워크 ───────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# ── 데이터 처리 ───────────────────────────────────────────────────────
pandas>=2.2.0
pyarrow>=15.0.0        # parquet 읽기/쓰기
networkx>=3.3           # 그래프 엔진

# ── HTTP / 환경변수 ───────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
